+++ controllers/comment_controller.js
@@ -4,15 +4,20 @@ const Blog = require('../models/blog');
 module.exports.create = async function(req,res){
     const {content,blogId} = req.body;
     try {
-        const blog = await Blog.findById(blogId);
-        const comment = await Comment.create({
-            content: content,
-            createdBy: req.user._id,
-            associatedBlog: blogId
-        });
-        blog.comment.push(comment)
-        blog.save();
-        return res.redirect(`back`)
+        if(req.user){
+            const blog = await Blog.findById(blogId);
+            const comment = await Comment.create({
+                content: content,
+                createdBy: req.user._id,
+                associatedBlog: blogId
+            });
+            blog.comment.push(comment)
+            blog.save();
+            return res.redirect(`back`)
+        }
+        else{
+            return res.redirect('/user/login')
+        }
     } catch (error) {
         console.log(`error in finding Blog`);
         return res.end(`404 Not Found`)
@@ -20,8 +25,11 @@ module.exports.create = async function(req,res){
 }
 module.exports.destroy = async function(req,res){
     try {
-        const commentId = req.params.id;
-        await Comment.findByIdAndDelete(commentId);
+        if(req.user){
+            const commentId = req.params.id;
+            await Comment.findByIdAndDelete(commentId);
+            return res.redirect('back')
+        }
         return res.redirect('back')
     } catch (error) {
         console.log("Error in deleting the Comment")
